REF: move Series-specific methods from NDFrame
